misc: modules for testing needs were moved to separate package
