+++ tests/functional-tests/read-write-tests/test_offset.py
@@ -5,11 +5,11 @@ import elliptics
 
 from hamcrest import assert_that, equal_to, has_property, has_length, is_
 
-import elliptics_testhelper as et
-from elliptics_testhelper import key_and_data, timestamp, user_flags, client, nodes
+import test_helper.elliptics_testhelper as et
+import test_helper.utils as utils
+from test_helper.elliptics_testhelper import key_and_data, timestamp, user_flags, client, nodes
+from test_helper.matchers import elliptics_result_with
 
-import utils
-from matchers import elliptics_result_with
 from testcases import offset_and_chunksize_types_list, \
                       write_offset_type_and_overriding_list
