+++ tests/system-tests/recovery-tests/test_recovery.py
@@ -11,12 +11,11 @@ from hamcrest import assert_that, raises, calling, equal_to
 
 import elliptics
 
-import elliptics_testhelper as et
-
-from elliptics_testhelper import nodes
-from utils import MB, get_key_and_data
-from matchers import hasitems
-from logging_tests import logger
+import test_helper.elliptics_testhelper as et
+from test_helper.elliptics_testhelper import nodes
+from test_helper.utils import MB, get_key_and_data
+from test_helper.matchers import hasitems
+from test_helper.logging_tests import logger
 
 @pytest.fixture(scope='module')
 def client(pytestconfig, nodes):
